Fix file uploads that don't use arrays.
This patch fixes a regression added by the previous patch which enabled
uploading files to arrays, using square brackets in form field names.
The eval() code used to parse the complex array syntax had a syntax error
if the name of the file did not have square brackets.  With this patch
eval() is only used when there are some.