use stream_select() prior to stream_socket_accept() to workaround non-blocking accept() call on ARM machines
It seems that accept() is ignoring the "infinite timeout" parameter (0
or -1) on ARM. This makes it return instantly without an actual
connection. This kills the worker, which is restarted immediately. We
could just wait for let's say 1 hour and live with the fact that our
workers will be restarted again and again. But this fix is better.
We cannot use the (functioning!) socket API directly (without stream
support), because more stream functions are used later on that would
have to be replaced as well.

We replaced the stream_get_line() with fgets() to work around another
issue: When selecting and accepting the connection, the s_g_l() would
block for the default_socket_timeout duration before reading the request
and handling it.